requirements.txt
streamlit
pymongo
pymongo[srv]
google-generativeai
python-docx
dnspython
